requirements.txt
discord.py>=2.3.0
anthropic>=0.39.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytz>=2024.1
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
pytest>=8.0.0
